refactor: extract looksLikeMeetingInvite to calendar-utils
