+++ src/lib/calendar-utils.ts
@@ -328,3 +328,19 @@ export const EVENT_COLORS = {
     text: "var(--green)",
   },
 } as const;
+
+// ─── Description helpers ─────────────────────────────────────────
+
+/** Check if an event description looks like meeting invite boilerplate. */
+export function looksLikeMeetingInvite(desc: string): boolean {
+  const lower = desc.toLowerCase();
+  const signals = [
+    "join zoom", "zoom.us/j/", "meeting id:", "dial by your location",
+    "join microsoft teams", "teams.microsoft.com", "click here to join",
+    "meet.google.com/", "join with google meet",
+    "zoho.com/meeting", "join meeting",
+    "passcode:", "one tap mobile", "dial-in", "phone number",
+  ];
+  const hits = signals.filter((s) => lower.includes(s)).length;
+  return hits >= 2;
+}
