+++ src/lib/components/calendar/EventSidePanel.svelte
@@ -11,6 +11,7 @@
     platformLabel,
     formatTimeRange,
     formatDayHeader,
+    looksLikeMeetingInvite,
   } from "../../calendar-utils";
   import BriefingPanel from "../BriefingPanel.svelte";
   import type { CalendarEvent } from "../../tauri";
@@ -23,19 +24,6 @@
   }
 
   let { event, matchedId, onClose, onOpenPopover }: Props = $props();
-
-  function looksLikeMeetingInvite(desc: string): boolean {
-    const lower = desc.toLowerCase();
-    const signals = [
-      "join zoom", "zoom.us/j/", "meeting id:", "dial by your location",
-      "join microsoft teams", "teams.microsoft.com", "click here to join",
-      "meet.google.com/", "join with google meet",
-      "zoho.com/meeting", "join meeting",
-      "passcode:", "one tap mobile", "dial-in", "phone number",
-    ];
-    const hits = signals.filter((s) => lower.includes(s)).length;
-    return hits >= 2;
-  }
 
   // Participants filtered to exclude current user
   let visibleParticipants = $derived.by(() => {
